Add MIT License to project with @grusad's approval
From the email conversation we just had:

"Well, haven't done anything with the channel in years, everything is more or less up for grabs. So you have my permission to use anything and/or everything however you like. You can add the MIT license to your fork.

You can refer to this conversation if needed!"